+++ app/assets/javascripts/other/readiness_statuses.js
@@ -1,5 +1,5 @@
 $(document).ready(function () {
-  $("[name='readiness_status[ready]']").click(function () {
+  $("[name='person[ready]']").click(function () {
     var showReason = $(this).val() === "false";
     var $reason    = $("#unreadiness_reason_form_group");
     $reason.toggle(showReason);